+++ src/cubicmultispline/__init__.py
@@ -6,6 +6,13 @@ from .Spline1D import Spline1D
 from .Spline import Spline
 
 __all__ = [
-    "Spline1D", 
+    "Spline1D",
     "Spline"
 ]
+
+try:
+    from .TorchSpline1D import TorchSpline1D
+    from .TorchSpline import TorchSpline
+    __all__ += ["TorchSpline1D", "TorchSpline"]
+except ImportError:
+    pass